Move `ResetIdentity` to shared steps.
