+++ packages/sdk/shell/src/panels/JoinPanel/JoinPanelProps.ts
@@ -8,8 +8,8 @@ import type { Identity } from '@dxos/react-client/halo';
 import type { Invitation, AuthenticatingInvitationObservable, InvitationResult } from '@dxos/react-client/invitations';
 
 import { type JoinSend } from './joinMachine';
-import { type ResetIdentityProps, type IdentityInputProps } from './steps';
-import { type StepProps } from '../../steps';
+import { type IdentityInputProps } from './steps';
+import { type ResetIdentityProps, type StepProps } from '../../steps';
 import { type FailReason } from '../../types';
 
 export type JoinPanelMode = 'default' | 'halo-only';
