+++ packages/sdk/shell/src/translations/locales/en-US.ts
@@ -73,7 +73,7 @@ export const os = {
   'choose add device label': 'Add device',
   'choose devices label': 'Manage devices',
   'choose profile label': 'Edit profile',
-  'choose sign out label': 'Sign out',
+  'choose sign out label': 'Reset storage',
   'empty space members message': 'There is nobody in this space yet.',
   'back to all invitations label': 'Back to all invitations',
   'invitation heading': 'Invitation',
@@ -104,7 +104,7 @@ export const os = {
   'sign out chooser title': 'DANGER ZONE',
   'sign out chooser message': 'Proceeding with either action below will erase all data on this device.',
   'join new identity label': 'Join a new identity',
-  'reset device label': 'Reset device',
+  'reset device label': 'Reset storage',
   'reset device confirm message': 'Are you sure you want to reset your device? All data will be lost.',
   'emoji and color label': 'Emoji and color',
   'clear label': 'Use default',
